+++ docker/bind9/zones/not.loaded.hosts
@@ -1,0 +1,9 @@
+$ttl 38400
+not.loaded.	IN	SOA	172.17.42.1. admin.test.com. (
+			1439234395
+			10800
+			3600
+			604800
+			38400 )
+not.loaded.      IN  NS  172.17.42.1.
+foo              IN  A   1.1.1.1
