+++ docker/conf/application.conf
@@ -132,6 +132,10 @@ vinyldns {
   # the max number of changes in a single batch change.  Change carefully as this has performance
   # implications
   batch-change-limit = 1000
+
+  manual-batch-review-enabled = true
+
+  scheduled-changes-enabled = true
 }
 
 # Akka settings, these should not need to be modified unless you know akka http really well.
@@ -157,7 +161,3 @@ akka.http {
     illegal-header-warnings = on
   }
 }
-
-manual-batch-review-enabled = true
-
-scheduled-changes-enabled = true
